+++ app/views/referrals/organisation/check_answers/edit.html.erb
@@ -3,7 +3,7 @@
 
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds">
-    <%= form_with model: @organisation_form, url: referral_organisation_path(current_referral), method: :patch do |f| %>
+    <%= form_with model: @organisation_form, url: referral_organisation_check_answers_path(current_referral), method: :patch do |f| %>
       <%= f.govuk_error_summary %>
 
       <h1 class="govuk-heading-l">
